refactor: move Map and RadioCard out of feature components folder
